Fixed a problem that can delete others' posts and comments

+++ app/Http/Controllers/CommentsController.php
@@ -26,7 +26,12 @@ class CommentsController extends Controller {
   }
 
   public function destroy(Tweet $tweet, Comment $comment) {
-    $comment->delete();
-    return redirect()->back();
+    if (Auth::check() && Auth::id() == $comment->user_id) {
+      $comment->delete();
+      return redirect()->back();
+    } else {
+      return redirect()->back();
+    }
+
   }
 }
